fix: allow unauthenticated users
I had assumed that users needed an access token to use the GraphQL API
but the API seems to work without an access token. The extension now
allows users to use the extension without an access token. We still need
to handle the case when users have provided an invalid access token.